+++ ext/js/app/popup-window.js
@@ -34,7 +34,7 @@ export class PopupWindow extends EventDispatcher {
         frameId
     }) {
         super();
-        /** @type {import('../yomitan.js').Application} */
+        /** @type {import('../application.js').Application} */
         this._application = application;
         /** @type {string} */
         this._id = id;
