+++ types/ext/popup.d.ts
@@ -19,7 +19,7 @@ import type {Popup} from '../../ext/js/app/popup';
 import type {PopupProxy} from '../../ext/js/app/popup-proxy';
 import type {PopupWindow} from '../../ext/js/app/popup-window';
 import type {FrameOffsetForwarder} from '../../ext/js/comm/frame-offset-forwarder';
-import type {Application} from '../../ext/js/yomitan';
+import type {Application} from '../../ext/js/application';
 import type * as DocumentUtil from './document-util';
 import type * as Settings from './settings';
 import type {EventNames, EventArgument as BaseEventArgument} from './core';
